Re-add get all for lang and dt only + fix tests with input patch
