+++ tests/cli/test_get.py
@@ -47,28 +47,29 @@ class TestGetData(unittest.TestCase):
 
     # MARK: All Data
 
+    @patch("scribe_data.cli.get.query_data")
+    @patch("builtins.input", lambda _: "N")  # don't use dump
+    def test_get_all_data_types_for_language(self, mock_query_data):
+        get_data(all=True, language="English")
+        mock_query_data.assert_called_once_with(
+            languages=["English"],
+            data_type=None,
+            output_dir="scribe_data_json_export",
+            overwrite=False,
+        )
+
+    @patch("scribe_data.cli.get.query_data")
+    @patch("builtins.input", lambda _: "N")  # don't use dump
+    def test_get_all_languages_for_data_type(self, mock_query_data):
+        get_data(all=True, data_type="nouns")
+        mock_query_data.assert_called_once_with(
+            languages=None,
+            data_type=["nouns"],
+            output_dir="scribe_data_json_export",
+            overwrite=False,
+        )
+
     # Note: Wikidata dumps are required for extracting all data.
-
-    # @patch("scribe_data.cli.get.query_data")
-    # def test_get_all_data_types_for_language(self, mock_query_data):
-    #     get_data(all=True, language="English")
-    #     mock_query_data.assert_called_once_with(
-    #         languages=["English"],
-    #         data_type=None,
-    #         output_dir="scribe_data_json_export",
-    #         overwrite=False,
-    #     )
-
-    # @patch("scribe_data.cli.get.query_data")
-    # def test_get_all_languages_for_data_type(self, mock_query_data):
-    #     get_data(all=True, data_type="nouns")
-    #     mock_query_data.assert_called_once_with(
-    #         languages=None,
-    #         data_type=["nouns"],
-    #         output_dir="scribe_data_json_export",
-    #         overwrite=False,
-    #     )
-
     # @patch("scribe_data.cli.get.query_data")
     # def test_get_all_languages_and_data_types(self, mock_query_data):
     #     get_data(all=True, output_dir="./test_output")
